PR Comments: 1. Merge ComparatorCursor and ComparatorCursorBase 2. Rename protobuf fields and use optional instead of required 3. Extract PlanSelector interface to top level 4. null byte[] for isEnd continuation 5. ThreadLocalRandom 6. Memoize string and hash suppliers
